Don't destroy the login keyboards when installing a new one
Because when installing a keyboard, kbdi is ran a `NT AUTHORITY/SYSTEM`,
targeting `HKCU` modifies `HKU/.DEFAULT` which contains the keyboard
list shown on the login screen, `HKEY_USERS\.DEFAULT\Keyboard
Layout\Preload` and because `NT AUTHORITY/SYSTEM` doesn't have any
keyboard configuration (it's not a user, why would it), we would end up
with that list being emptied and then the newly installed keyboard added
to it. Which in some case can be very problematic as you might not be
able to type your password.

By ignoring the `regenerate_given_registry` call when kbdi is called as
`NT AUTHORITY/SYSTEM`, we avoid that behavior. It even looks like
windows is nice with us and is adding keyboards to the login screen for
us.